Fixed computation of the ray-box intersections.
Even according to the link to the used method, the approach should only
work for axis-aligned boxes. But bodies::Box can be arbitrarily transformed.
So I instead strip rotation from both the box and the ray, which gives
the same results.

The old implementation would fail every time the untransformed min corner
ceased to be the min corner of the transformed box.

# Conflicts:
#	src/bodies.cpp